+++ source/commands/utils/sharedDangerfileArgs.ts
@@ -27,7 +27,7 @@ export interface SharedCLI extends program.CommanderStatic {
   failOnErrors: boolean
   /** Use GitHub Checks */
   useGithubChecks: boolean
-  /** Ignore Inline-Comments where the specified line was not changed */
+  /** Ignore invalid inline-comments, for instance a comment with a line that was not changed */
   ignoreInvalidInlineComments: boolean
 }
 
@@ -45,3 +45,8 @@ export default (command: any) =>
     .option("-u, --passURLForDSL", "[dev] Use a custom URL to send the Danger DSL into the sub-process")
     .option("-f, --failOnErrors", "Fail if there are fails in the danger report", false)
     .option("--use-github-checks", "Use GitHub Checks", false)
+    .option(
+      "-iiic, --ignoreInvalidInlineComments",
+      "Ignore invalid inline-comments, for instance a comment with a line that was not changed",
+      false
+    )